@@ -1,56 +1,74 @@
 <?php
+declare(strict_types=1);
 
 namespace Faker;
 
+use DateTime;
+use InvalidArgumentException;
+use ReflectionMethod;
+use ReflectionObject;
+
+use Faker\Api\FakerGeneratorInterface;
+use Faker\Provider\Base as BaseProvider;
+
+use function array_reverse;
+use function get_class;
+use function var_export;
+use function join;
+
 class Documentor
 {
-    protected $generator;
 
     /**
-     * @param Generator $generator
+     * @param FakerGeneratorInterface $generator
      */
-    public function __construct(Generator $generator)
-    {
-        $this->generator = $generator;
-    }
+    public function __construct(
+        protected FakerGeneratorInterface $generator
+    ) {}
 
     /**
      * @return array
      */
-    public function getFormatters()
+    public function getFormatters(): array
     {
-        $formatters = array();
+        $formatters = [];
         $providers = array_reverse($this->generator->getProviders());
-        $providers[]= new Provider\Base($this->generator);
+        $providers[] = new BaseProvider($this->generator);
         foreach ($providers as $provider) {
             $providerClass = get_class($provider);
-            $formatters[$providerClass] = array();
-            $refl = new \ReflectionObject($provider);
-            foreach ($refl->getMethods(\ReflectionMethod::IS_PUBLIC) as $reflmethod) {
-                if ($reflmethod->getDeclaringClass()->getName() == 'Faker\Provider\Base' && $providerClass != 'Faker\Provider\Base') {
+            $formatters[$providerClass] = [];
+            $reflectionObject = new ReflectionObject($provider);
+            foreach ($reflectionObject->getMethods(ReflectionMethod::IS_PUBLIC) as $reflectionMethod) {
+                if ($reflectionMethod->getDeclaringClass()->getName() == 'Faker\Provider\Base' && $providerClass != 'Faker\Provider\Base') {
                     continue;
                 }
-                $methodName = $reflmethod->name;
-                if ($reflmethod->isConstructor()) {
+
+                $methodName = $reflectionMethod->name;
+                if ($reflectionMethod->isConstructor()) {
                     continue;
                 }
-                $parameters = array();
-                foreach ($reflmethod->getParameters() as $reflparameter) {
-                    $parameter = '$'. $reflparameter->getName();
-                    if ($reflparameter->isDefaultValueAvailable()) {
-                        $parameter .= ' = ' . var_export($reflparameter->getDefaultValue(), true);
+
+                $parameters = [];
+                foreach ($reflectionMethod->getParameters() as $reflectionParameter) {
+                    $parameter = '$'. $reflectionParameter->getName();
+                    if ($reflectionParameter->isDefaultValueAvailable()) {
+                        $parameter .= ' = ' . var_export($reflectionParameter->getDefaultValue(), true);
                     }
+
                     $parameters []= $parameter;
                 }
+
                 $parameters = $parameters ? '('. join(', ', $parameters) . ')' : '';
+
                 try {
                     $example = $this->generator->format($methodName);
-                } catch (\InvalidArgumentException $e) {
+                } catch (InvalidArgumentException $e) {
                     $example = '';
                 }
+
                 if (is_array($example)) {
                     $example = "array('". join("', '", $example) . "')";
-                } elseif ($example instanceof \DateTime) {
+                } elseif ($example instanceof DateTime) {
                     $example = "DateTime('" . $example->format('Y-m-d H:i:s') . "')";
                 } elseif ($example instanceof Generator || $example instanceof UniqueGenerator) { // modifier
                     $example = '';
