@@ -1,56 +1,96 @@
 <?php
+declare(strict_types=1);
 
 namespace Faker\Provider;
 
 class Address extends Base
 {
-    protected static $citySuffix = array('Ville');
-    protected static $streetSuffix = array('Street');
-    protected static $cityFormats = array(
-        '{{firstName}}{{citySuffix}}',
-    );
-    protected static $streetNameFormats = array(
-        '{{lastName}} {{streetSuffix}}'
-    );
-    protected static $streetAddressFormats = array(
-        '{{buildingNumber}} {{streetName}}'
-    );
-    protected static $addressFormats = array(
-        '{{streetAddress}} {{postcode}} {{city}}',
-    );
 
-    protected static $buildingNumber = array('%#');
-    protected static $postcode = array('#####');
-    protected static $country = array();
+    /**
+     * @var string[]
+     */
+    protected static array $citySuffix = ['Ville'];
+
+    /**
+     * @var string[]
+     */
+    protected static array $streetSuffix = ['Street'];
+
+    /**
+     * @var string[]
+     */
+    protected static array $cityFormats = [
+        '{{firstName}}{{citySuffix}}'
+    ];
+
+    /**
+     * @var string[]
+     */
+    protected static array $streetNameFormats = [
+        '{{lastName}} {{streetSuffix}}'
+    ];
+
+    /**
+     * @var string[]
+     */
+    protected static array $streetAddressFormats = [
+        '{{buildingNumber}} {{streetName}}'
+    ];
+
+    /**
+     * @var string[]
+     */
+    protected static array $addressFormats = [
+        '{{streetAddress}} {{postcode}} {{city}}',
+    ];
+
+    /**
+     * @var string[]
+     */
+    protected static array $buildingNumber = ['%#'];
+
+    /**
+     * @var string[]
+     */
+    protected static array $postcode = ['#####'];
+
+    /**
+     * @var string[]
+     */
+    protected static array $country = [];
 
     /**
      * @example 'town'
+     * @return string
      */
-    public static function citySuffix()
+    public static function citySuffix(): string
     {
         return static::randomElement(static::$citySuffix);
     }
 
     /**
      * @example 'Avenue'
+     * @return string
      */
-    public static function streetSuffix()
+    public static function streetSuffix(): string
     {
         return static::randomElement(static::$streetSuffix);
     }
 
     /**
      * @example '791'
+     * @return string
      */
-    public static function buildingNumber()
+    public static function buildingNumber(): string
     {
         return static::numerify(static::randomElement(static::$buildingNumber));
     }
 
     /**
      * @example 'Sashabury'
+     * @return string
      */
-    public function city()
+    public function city(): string
     {
         $format = static::randomElement(static::$cityFormats);
 
@@ -59,8 +99,9 @@ class Address extends Base
 
     /**
      * @example 'Crist Parks'
+     * @return string
      */
-    public function streetName()
+    public function streetName(): string
     {
         $format = static::randomElement(static::$streetNameFormats);
 
@@ -69,8 +110,9 @@ class Address extends Base
 
     /**
      * @example '791 Crist Parks'
+     * @return string
      */
-    public function streetAddress()
+    public function streetAddress(): string
     {
         $format = static::randomElement(static::$streetAddressFormats);
 
@@ -79,16 +121,18 @@ class Address extends Base
 
     /**
      * @example 86039-9874
+     * @return string
      */
-    public static function postcode()
+    public static function postcode(): string
     {
         return static::toUpper(static::bothify(static::randomElement(static::$postcode)));
     }
 
     /**
      * @example '791 Crist Parks, Sashabury, IL 86039-9874'
+     * @return string
      */
-    public function address()
+    public function address(): string
     {
         $format = static::randomElement(static::$addressFormats);
 
@@ -97,8 +141,9 @@ class Address extends Base
 
     /**
      * @example 'Japan'
+     * @return string
      */
-    public static function country()
+    public static function country(): string
     {
         return static::randomElement(static::$country);
     }
@@ -109,7 +154,7 @@ class Address extends Base
      * @param float|int $max
      * @return float Uses signed degrees format (returns a float number between -90 and 90)
      */
-    public static function latitude($min = -90, $max = 90)
+    public static function latitude(float|int $min = -90, float|int $max = 90): float
     {
         return static::randomFloat(6, $min, $max);
     }
@@ -120,16 +165,16 @@ class Address extends Base
      * @param float|int $max
      * @return float Uses signed degrees format (returns a float number between -180 and 180)
      */
-    public static function longitude($min = -180, $max = 180)
+    public static function longitude(float|int $min = -180, float|int $max = 180): float
     {
         return static::randomFloat(6, $min, $max);
     }
 
     /**
      * @example array('77.147489', '86.211205')
-     * @return array | latitude, longitude
+     * @return array  latitude, longitude
      */
-    public static function localCoordinates()
+    public static function localCoordinates(): array
     {
         return array(
             'latitude' => static::latitude(),
