@@ -1,19 +1,31 @@
 <?php
+declare(strict_types=1);
 
 namespace Faker\Provider;
+
+use function str_pad;
+use function dechex;
+use function hexdec;
+use function mt_rand;
+use function substr;
+use function implode;
 
 /**
  * @author lsv
  */
 class Color extends Base
 {
-    protected static $safeColorNames = array(
+
+    /**
+     * @var string[]
+     */
+    protected static array $safeColorNames = [
         'black', 'maroon', 'green', 'navy', 'olive',
         'purple', 'teal', 'lime', 'blue', 'silver',
         'gray', 'yellow', 'fuchsia', 'aqua', 'white'
-    );
+    ];
 
-    protected static $allColorNames = array(
+    protected static array $allColorNames = [
         'AliceBlue', 'AntiqueWhite', 'Aqua', 'Aquamarine',
         'Azure', 'Beige', 'Bisque', 'Black', 'BlanchedAlmond',
         'Blue', 'BlueViolet', 'Brown', 'BurlyWood', 'CadetBlue',
@@ -40,20 +52,22 @@ class Color extends Base
         'SandyBrown', 'SeaGreen', 'SeaShell', 'Sienna', 'Silver', 'SkyBlue', 'SlateBlue',
         'SlateGray', 'Snow', 'SpringGreen', 'SteelBlue', 'Tan', 'Teal', 'Thistle', 'Tomato',
         'Turquoise', 'Violet', 'Wheat', 'White', 'WhiteSmoke', 'Yellow', 'YellowGreen'
-    );
+    ];
 
     /**
      * @example '#fa3cc2'
+     * @return string
      */
-    public static function hexColor()
+    public static function hexColor(): string
     {
         return '#' . str_pad(dechex(mt_rand(1, 16777215)), 6, '0', STR_PAD_LEFT);
     }
 
     /**
      * @example '#ff0044'
+     * @return string
      */
-    public static function safeHexColor()
+    public static function safeHexColor(): string
     {
         $color = str_pad(dechex(mt_rand(0, 255)), 3, '0', STR_PAD_LEFT);
 
@@ -62,54 +76,60 @@ class Color extends Base
 
     /**
      * @example 'array(0,255,122)'
+     * @return string[]
      */
-    public static function rgbColorAsArray()
+    public static function rgbColorAsArray(): array
     {
         $color = static::hexColor();
 
-        return array(
+        return [
             hexdec(substr($color, 1, 2)),
             hexdec(substr($color, 3, 2)),
             hexdec(substr($color, 5, 2))
-        );
+        ];
     }
 
     /**
      * @example '0,255,122'
+     * @return string
      */
-    public static function rgbColor()
+    public static function rgbColor(): string
     {
         return implode(',', static::rgbColorAsArray());
     }
 
     /**
      * @example 'rgb(0,255,122)'
+     * @return string
      */
-    public static function rgbCssColor()
+    public static function rgbCssColor(): string
     {
         return 'rgb(' . static::rgbColor() . ')';
     }
 
     /**
      * @example 'rgba(0,255,122,0.8)'
+     * @return string
      */
-    public static function rgbaCssColor()
+    public static function rgbaCssColor(): string
     {
         return 'rgba(' . static::rgbColor() . ',' . static::randomFloat(1, 0, 1) . ')';
     }
 
     /**
      * @example 'blue'
+     * @return string
      */
-    public static function safeColorName()
+    public static function safeColorName(): string
     {
         return static::randomElement(static::$safeColorNames);
     }
 
     /**
      * @example 'NavajoWhite'
+     * @return string
      */
-    public static function colorName()
+    public static function colorName(): string
     {
         return static::randomElement(static::$allColorNames);
     }
@@ -118,7 +138,7 @@ class Color extends Base
      * @example '340,50,20'
      * @return string
      */
-    public static function hslColor()
+    public static function hslColor(): string
     {
         return sprintf(
             '%s,%s,%s',
@@ -130,14 +150,14 @@ class Color extends Base
 
     /**
      * @example array(340, 50, 20)
-     * @return array
+     * @return int[]
      */
-    public static function hslColorAsArray()
+    public static function hslColorAsArray(): array
     {
-        return array(
+        return [
             static::numberBetween(0, 360),
             static::numberBetween(0, 100),
             static::numberBetween(0, 100)
-        );
+        ];
     }
 }
