@@ -1,8 +1,29 @@
 <?php
+declare(strict_types=1);
+
 namespace Faker\Provider;
+
+use InvalidArgumentException;
+
+use function is_dir;
+use function is_file;
+use function is_readable;
+use function is_array;
+use function pathinfo;
+use function copy;
+use function scandir;
+use function sprintf;
+use function array_filter;
+use function array_values;
+use function array_diff;
 
 class File extends Base
 {
+
+    /**
+     * @var string
+     */
+    protected const TMP_DIR = '/tmp';
 
     /**
      * MIME types from the apache.org file. Some types are truncated.
@@ -10,7 +31,7 @@ class File extends Base
      * @var array Map of MIME types => file extension(s)
      * @link http://svn.apache.org/repos/asf/httpd/httpd/trunk/docs/conf/mime.types
      */
-    protected static $mimeTypes = array(
+    protected static array $mimeTypes = [
         'application/atom+xml'                                                      => 'atom',
         'application/ecmascript'                                                    => 'ecma',
         'application/emma+xml'                                                      => 'emma',
@@ -25,8 +46,8 @@ class File extends Base
         'application/mets+xml'                                                      => 'mets',
         'application/mods+xml'                                                      => 'mods',
         'application/mp4'                                                           => 'mp4s',
-        'application/msword'                                                        => array('doc', 'dot'),
-        'application/octet-stream'                                                  => array(
+        'application/msword'                                                        => ['doc', 'dot'],
+        'application/octet-stream'                                                  => [
             'bin',
             'dms',
             'lrf',
@@ -39,16 +60,16 @@ class File extends Base
             'dump',
             'elc',
             'deploy'
-        ),
+        ],
         'application/ogg'                                                           => 'ogx',
         'application/omdoc+xml'                                                     => 'omdoc',
         'application/pdf'                                                           => 'pdf',
         'application/pgp-encrypted'                                                 => 'pgp',
-        'application/pgp-signature'                                                 => array('asc', 'sig'),
+        'application/pgp-signature'                                                 => ['asc', 'sig'],
         'application/pkix-pkipath'                                                  => 'pkipath',
         'application/pkixcmp'                                                       => 'pki',
         'application/pls+xml'                                                       => 'pls',
-        'application/postscript'                                                    => array('ai', 'eps', 'ps'),
+        'application/postscript'                                                    => ['ai', 'eps', 'ps'],
         'application/pskc+xml'                                                      => 'pskcxml',
         'application/rdf+xml'                                                       => 'rdf',
         'application/reginfo+xml'                                                   => 'rif',
@@ -61,10 +82,10 @@ class File extends Base
         'application/vnd.ahead.space'                                               => 'ahead',
         'application/vnd.dart'                                                      => 'dart',
         'application/vnd.data-vision.rdz'                                           => 'rdz',
-        'application/vnd.dece.data'                                                 => array('uvf', 'uvvf', 'uvd', 'uvvd'),
-        'application/vnd.dece.ttml+xml'                                             => array('uvt', 'uvvt'),
-        'application/vnd.dece.unspecified'                                          => array('uvx', 'uvvx'),
-        'application/vnd.dece.zip'                                                  => array('uvz', 'uvvz'),
+        'application/vnd.dece.data'                                                 => ['uvf', 'uvvf', 'uvd', 'uvvd'],
+        'application/vnd.dece.ttml+xml'                                             => ['uvt', 'uvvt'],
+        'application/vnd.dece.unspecified'                                          => ['uvx', 'uvvx'],
+        'application/vnd.dece.zip'                                                  => ['uvz', 'uvvz'],
         'application/vnd.denovo.fcselayout-link'                                    => 'fe_launch',
         'application/vnd.dna'                                                       => 'dna',
         'application/vnd.dolby.mlp'                                                 => 'mlp',
@@ -85,14 +106,14 @@ class File extends Base
         'application/vnd.ezpix-package'                                             => 'ez3',
         'application/vnd.fdf'                                                       => 'fdf',
         'application/vnd.fdsn.mseed'                                                => 'mseed',
-        'application/vnd.fdsn.seed'                                                 => array('seed', 'dataless'),
+        'application/vnd.fdsn.seed'                                                 => ['seed', 'dataless'],
         'application/vnd.flographit'                                                => 'gph',
         'application/vnd.fluxtime.clip'                                             => 'ftc',
         'application/vnd.hal+xml'                                                   => 'hal',
         'application/vnd.hydrostatix.sof-data'                                      => 'sfd-hdstx',
         'application/vnd.ibm.minipay'                                               => 'mpy',
         'application/vnd.ibm.secure-container'                                      => 'sc',
-        'application/vnd.iccprofile'                                                => array('icc', 'icm'),
+        'application/vnd.iccprofile'                                                => ['icc', 'icm'],
         'application/vnd.igloader'                                                  => 'igl',
         'application/vnd.immervision-ivp'                                           => 'ivp',
         'application/vnd.kde.karbon'                                                => 'karbon',
@@ -100,13 +121,13 @@ class File extends Base
         'application/vnd.kde.kformula'                                              => 'kfo',
         'application/vnd.kde.kivio'                                                 => 'flw',
         'application/vnd.kde.kontour'                                               => 'kon',
-        'application/vnd.kde.kpresenter'                                            => array('kpr', 'kpt'),
+        'application/vnd.kde.kpresenter'                                            => ['kpr', 'kpt'],
         'application/vnd.kde.kspread'                                               => 'ksp',
-        'application/vnd.kde.kword'                                                 => array('kwd', 'kwt'),
+        'application/vnd.kde.kword'                                                 => ['kwd', 'kwt'],
         'application/vnd.kenameaapp'                                                => 'htke',
         'application/vnd.kidspiration'                                              => 'kia',
-        'application/vnd.kinar'                                                     => array('kne', 'knp'),
-        'application/vnd.koan'                                                      => array('skp', 'skd', 'skt', 'skm'),
+        'application/vnd.kinar'                                                     => ['kne', 'knp'],
+        'application/vnd.koan'                                                      => ['skp', 'skd', 'skt', 'skm'],
         'application/vnd.kodak-descriptor'                                          => 'sse',
         'application/vnd.las.las+xml'                                               => 'lasxml',
         'application/vnd.llamagraphics.life-balance.desktop'                        => 'lbd',
@@ -120,14 +141,14 @@ class File extends Base
         'application/vnd.mozilla.xul+xml'                                           => 'xul',
         'application/vnd.ms-artgalry'                                               => 'cil',
         'application/vnd.ms-cab-compressed'                                         => 'cab',
-        'application/vnd.ms-excel'                                                  => array(
+        'application/vnd.ms-excel'                                                  => [
             'xls',
             'xlm',
             'xla',
             'xlc',
             'xlt',
             'xlw'
-        ),
+        ],
         'application/vnd.ms-excel.addin.macroenabled.12'                            => 'xlam',
         'application/vnd.ms-excel.sheet.binary.macroenabled.12'                     => 'xlsb',
         'application/vnd.ms-excel.sheet.macroenabled.12'                            => 'xlsm',
@@ -139,16 +160,16 @@ class File extends Base
         'application/vnd.ms-officetheme'                                            => 'thmx',
         'application/vnd.ms-pki.seccat'                                             => 'cat',
         'application/vnd.ms-pki.stl'                                                => 'stl',
-        'application/vnd.ms-powerpoint'                                             => array('ppt', 'pps', 'pot'),
+        'application/vnd.ms-powerpoint'                                             => ['ppt', 'pps', 'pot'],
         'application/vnd.ms-powerpoint.addin.macroenabled.12'                       => 'ppam',
         'application/vnd.ms-powerpoint.presentation.macroenabled.12'                => 'pptm',
         'application/vnd.ms-powerpoint.slide.macroenabled.12'                       => 'sldm',
         'application/vnd.ms-powerpoint.slideshow.macroenabled.12'                   => 'ppsm',
         'application/vnd.ms-powerpoint.template.macroenabled.12'                    => 'potm',
-        'application/vnd.ms-project'                                                => array('mpp', 'mpt'),
+        'application/vnd.ms-project'                                                => ['mpp', 'mpt'],
         'application/vnd.ms-word.document.macroenabled.12'                          => 'docm',
         'application/vnd.ms-word.template.macroenabled.12'                          => 'dotm',
-        'application/vnd.ms-works'                                                  => array('wps', 'wks', 'wcm', 'wdb'),
+        'application/vnd.ms-works'                                                  => ['wps', 'wks', 'wcm', 'wdb'],
         'application/vnd.ms-wpl'                                                    => 'wpl',
         'application/vnd.ms-xpsdocument'                                            => 'xps',
         'application/vnd.mseq'                                                      => 'mseq',
@@ -182,14 +203,14 @@ class File extends Base
         'application/vnd.openxmlformats-officedocument.wordprocessingml.document'   => 'docx',
         'application/vnd.openxmlformats-officedocument.wordprocessingml.template'   => 'dotx',
         'application/vnd.pvi.ptid1'                                                 => 'ptid',
-        'application/vnd.quark.quarkxpress'                                         => array(
+        'application/vnd.quark.quarkxpress'                                         => [
             'qxd',
             'qxt',
             'qwd',
             'qwt',
             'qxl',
             'qxb'
-        ),
+        ],
         'application/vnd.realvnc.bed'                                               => 'bed',
         'application/vnd.recordare.musicxml'                                        => 'mxl',
         'application/vnd.recordare.musicxml+xml'                                    => 'musicxml',
@@ -207,7 +228,7 @@ class File extends Base
         'application/vnd.shana.informed.formtemplate'                               => 'itp',
         'application/vnd.shana.informed.interchange'                                => 'iif',
         'application/vnd.shana.informed.package'                                    => 'ipk',
-        'application/vnd.simtech-mindmapper'                                        => array('twd', 'twds'),
+        'application/vnd.simtech-mindmapper'                                        => ['twd', 'twds'],
         'application/vnd.smaf'                                                      => 'mmf',
         'application/vnd.stepmania.stepchart'                                       => 'sm',
         'application/vnd.sun.xml.calc'                                              => 'sxc',
@@ -220,25 +241,25 @@ class File extends Base
         'application/vnd.sun.xml.writer'                                            => 'sxw',
         'application/vnd.sun.xml.writer.global'                                     => 'sxg',
         'application/vnd.sun.xml.writer.template'                                   => 'stw',
-        'application/vnd.sus-calendar'                                              => array('sus', 'susp'),
+        'application/vnd.sus-calendar'                                              => ['sus', 'susp'],
         'application/vnd.svd'                                                       => 'svd',
-        'application/vnd.symbian.install'                                           => array('sis', 'sisx'),
+        'application/vnd.symbian.install'                                           => ['sis', 'sisx'],
         'application/vnd.syncml+xml'                                                => 'xsm',
         'application/vnd.syncml.dm+wbxml'                                           => 'bdm',
         'application/vnd.syncml.dm+xml'                                             => 'xdm',
         'application/vnd.tao.intent-module-archive'                                 => 'tao',
-        'application/vnd.tcpdump.pcap'                                              => array('pcap', 'cap', 'dmp'),
+        'application/vnd.tcpdump.pcap'                                              => ['pcap', 'cap', 'dmp'],
         'application/vnd.tmobile-livetv'                                            => 'tmo',
         'application/vnd.trid.tpt'                                                  => 'tpt',
         'application/vnd.triscape.mxs'                                              => 'mxs',
         'application/vnd.trueapp'                                                   => 'tra',
-        'application/vnd.ufdl'                                                      => array('ufd', 'ufdl'),
+        'application/vnd.ufdl'                                                      => ['ufd', 'ufdl'],
         'application/vnd.uiq.theme'                                                 => 'utz',
         'application/vnd.umajin'                                                    => 'umj',
         'application/vnd.unity'                                                     => 'unityweb',
         'application/vnd.uoml+xml'                                                  => 'uoml',
         'application/vnd.vcx'                                                       => 'vcx',
-        'application/vnd.visio'                                                     => array('vsd', 'vst', 'vss', 'vsw'),
+        'application/vnd.visio'                                                     => ['vsd', 'vst', 'vss', 'vsw'],
         'application/vnd.visionary'                                                 => 'vis',
         'application/vnd.vsf'                                                       => 'vsf',
         'application/vnd.wap.wbxml'                                                 => 'wbxml',
@@ -258,7 +279,7 @@ class File extends Base
         'application/wspolicy+xml'                                                  => 'wspolicy',
         'application/x-7z-compressed'                                               => '7z',
         'application/x-bittorrent'                                                  => 'torrent',
-        'application/x-blorb'                                                       => array('blb', 'blorb'),
+        'application/x-blorb'                                                       => ['blb', 'blorb'],
         'application/x-bzip'                                                        => 'bz',
         'application/x-cdlink'                                                      => 'vcd',
         'application/x-cfs-compressed'                                              => 'cfs',
@@ -267,9 +288,9 @@ class File extends Base
         'application/x-conference'                                                  => 'nsc',
         'application/x-cpio'                                                        => 'cpio',
         'application/x-csh'                                                         => 'csh',
-        'application/x-debian-package'                                              => array('deb', 'udeb'),
+        'application/x-debian-package'                                              => ['deb', 'udeb'],
         'application/x-dgc-compressed'                                              => 'dgc',
-        'application/x-director'                                                    => array(
+        'application/x-director'                                                    => [
             'dir',
             'dcr',
             'dxr',
@@ -279,9 +300,9 @@ class File extends Base
             'w3d',
             'fgd',
             'swa'
-        ),
-        'application/x-font-ttf'                                                    => array('ttf', 'ttc'),
-        'application/x-font-type1'                                                  => array('pfa', 'pfb', 'pfm', 'afm'),
+        ],
+        'application/x-font-ttf'                                                    => ['ttf', 'ttc'],
+        'application/x-font-type1'                                                  => ['pfa', 'pfb', 'pfm', 'afm'],
         'application/x-font-woff'                                                   => 'woff',
         'application/x-freearc'                                                     => 'arc',
         'application/x-futuresplash'                                                => 'spl',
@@ -295,9 +316,9 @@ class File extends Base
         'application/x-iso9660-image'                                               => 'iso',
         'application/x-java-jnlp-file'                                              => 'jnlp',
         'application/x-latex'                                                       => 'latex',
-        'application/x-lzh-compressed'                                              => array('lzh', 'lha'),
+        'application/x-lzh-compressed'                                              => ['lzh', 'lha'],
         'application/x-mie'                                                         => 'mie',
-        'application/x-mobipocket-ebook'                                            => array('prc', 'mobi'),
+        'application/x-mobipocket-ebook'                                            => ['prc', 'mobi'],
         'application/x-ms-application'                                              => 'application',
         'application/x-ms-shortcut'                                                 => 'lnk',
         'application/x-ms-wmd'                                                      => 'wmd',
@@ -307,13 +328,13 @@ class File extends Base
         'application/x-msbinder'                                                    => 'obd',
         'application/x-mscardfile'                                                  => 'crd',
         'application/x-msclip'                                                      => 'clp',
-        'application/x-msdownload'                                                  => array('exe', 'dll', 'com', 'bat', 'msi'),
-        'application/x-msmediaview'                                                 => array(
+        'application/x-msdownload'                                                  => ['exe', 'dll', 'com', 'bat', 'msi'],
+        'application/x-msmediaview'                                                 => [
             'mvb',
             'm13',
             'm14'
-        ),
-        'application/x-msmetafile'                                                  => array('wmf', 'wmz', 'emf', 'emz'),
+        ],
+        'application/x-msmetafile'                                                  => ['wmf', 'wmz', 'emf', 'emz'],
         'application/x-rar-compressed'                                              => 'rar',
         'application/x-research-info-systems'                                       => 'ris',
         'application/x-sh'                                                          => 'sh',
@@ -332,11 +353,11 @@ class File extends Base
         'application/x-tcl'                                                         => 'tcl',
         'application/x-tex'                                                         => 'tex',
         'application/x-tex-tfm'                                                     => 'tfm',
-        'application/x-texinfo'                                                     => array('texinfo', 'texi'),
+        'application/x-texinfo'                                                     => ['texinfo', 'texi'],
         'application/x-tgif'                                                        => 'obj',
         'application/x-ustar'                                                       => 'ustar',
         'application/x-wais-source'                                                 => 'src',
-        'application/x-x509-ca-cert'                                                => array('der', 'crt'),
+        'application/x-x509-ca-cert'                                                => ['der', 'crt'],
         'application/x-xfig'                                                        => 'fig',
         'application/x-xliff+xml'                                                   => 'xlf',
         'application/x-xpinstall'                                                   => 'xpi',
@@ -345,42 +366,42 @@ class File extends Base
         'application/xaml+xml'                                                      => 'xaml',
         'application/xcap-diff+xml'                                                 => 'xdf',
         'application/xenc+xml'                                                      => 'xenc',
-        'application/xhtml+xml'                                                     => array('xhtml', 'xht'),
-        'application/xml'                                                           => array('xml', 'xsl'),
+        'application/xhtml+xml'                                                     => ['xhtml', 'xht'],
+        'application/xml'                                                           => ['xml', 'xsl'],
         'application/xml-dtd'                                                       => 'dtd',
         'application/xop+xml'                                                       => 'xop',
         'application/xproc+xml'                                                     => 'xpl',
         'application/xslt+xml'                                                      => 'xslt',
         'application/xspf+xml'                                                      => 'xspf',
-        'application/xv+xml'                                                        => array('mxml', 'xhvml', 'xvml', 'xvm'),
+        'application/xv+xml'                                                        => ['mxml', 'xhvml', 'xvml', 'xvm'],
         'application/yang'                                                          => 'yang',
         'application/yin+xml'                                                       => 'yin',
         'application/zip'                                                           => 'zip',
         'audio/adpcm'                                                               => 'adp',
-        'audio/basic'                                                               => array('au', 'snd'),
-        'audio/midi'                                                                => array('mid', 'midi', 'kar', 'rmi'),
+        'audio/basic'                                                               => ['au', 'snd'],
+        'audio/midi'                                                                => ['mid', 'midi', 'kar', 'rmi'],
         'audio/mp4'                                                                 => 'mp4a',
-        'audio/mpeg'                                                                => array(
+        'audio/mpeg'                                                                => [
             'mpga',
             'mp2',
             'mp2a',
             'mp3',
             'm2a',
             'm3a'
-        ),
-        'audio/ogg'                                                                 => array('oga', 'ogg', 'spx'),
-        'audio/vnd.dece.audio'                                                      => array('uva', 'uvva'),
+        ],
+        'audio/ogg'                                                                 => ['oga', 'ogg', 'spx'],
+        'audio/vnd.dece.audio'                                                      => ['uva', 'uvva'],
         'audio/vnd.rip'                                                             => 'rip',
         'audio/webm'                                                                => 'weba',
         'audio/x-aac'                                                               => 'aac',
-        'audio/x-aiff'                                                              => array('aif', 'aiff', 'aifc'),
+        'audio/x-aiff'                                                              => ['aif', 'aiff', 'aifc'],
         'audio/x-caf'                                                               => 'caf',
         'audio/x-flac'                                                              => 'flac',
         'audio/x-matroska'                                                          => 'mka',
         'audio/x-mpegurl'                                                           => 'm3u',
         'audio/x-ms-wax'                                                            => 'wax',
         'audio/x-ms-wma'                                                            => 'wma',
-        'audio/x-pn-realaudio'                                                      => array('ram', 'ra'),
+        'audio/x-pn-realaudio'                                                      => ['ram', 'ra'],
         'audio/x-pn-realaudio-plugin'                                               => 'rmp',
         'audio/x-wav'                                                               => 'wav',
         'audio/xm'                                                                  => 'xm',
@@ -389,17 +410,17 @@ class File extends Base
         'image/g3fax'                                                               => 'g3',
         'image/gif'                                                                 => 'gif',
         'image/ief'                                                                 => 'ief',
-        'image/jpeg'                                                                => array('jpeg', 'jpg', 'jpe'),
+        'image/jpeg'                                                                => ['jpeg', 'jpg', 'jpe'],
         'image/ktx'                                                                 => 'ktx',
         'image/png'                                                                 => 'png',
         'image/prs.btif'                                                            => 'btif',
         'image/sgi'                                                                 => 'sgi',
-        'image/svg+xml'                                                             => array('svg', 'svgz'),
-        'image/tiff'                                                                => array('tiff', 'tif'),
+        'image/svg+xml'                                                             => ['svg', 'svgz'],
+        'image/tiff'                                                                => ['tiff', 'tif'],
         'image/vnd.adobe.photoshop'                                                 => 'psd',
-        'image/vnd.dece.graphic'                                                    => array('uvi', 'uvvi', 'uvg', 'uvvg'),
+        'image/vnd.dece.graphic'                                                    => ['uvi', 'uvvi', 'uvg', 'uvvg'],
         'image/vnd.dvb.subtitle'                                                    => 'sub',
-        'image/vnd.djvu'                                                            => array('djvu', 'djv'),
+        'image/vnd.djvu'                                                            => ['djvu', 'djv'],
         'image/vnd.dwg'                                                             => 'dwg',
         'image/vnd.dxf'                                                             => 'dxf',
         'image/vnd.fastbidsheet'                                                    => 'fbs',
@@ -416,11 +437,11 @@ class File extends Base
         'image/x-3ds'                                                               => '3ds',
         'image/x-cmu-raster'                                                        => 'ras',
         'image/x-cmx'                                                               => 'cmx',
-        'image/x-freehand'                                                          => array('fh', 'fhc', 'fh4', 'fh5', 'fh7'),
+        'image/x-freehand'                                                          => ['fh', 'fhc', 'fh4', 'fh5', 'fh7'],
         'image/x-icon'                                                              => 'ico',
         'image/x-mrsid-image'                                                       => 'sid',
         'image/x-pcx'                                                               => 'pcx',
-        'image/x-pict'                                                              => array('pic', 'pct'),
+        'image/x-pict'                                                              => ['pic', 'pct'],
         'image/x-portable-anymap'                                                   => 'pnm',
         'image/x-portable-bitmap'                                                   => 'pbm',
         'image/x-portable-graymap'                                                  => 'pgm',
@@ -430,26 +451,26 @@ class File extends Base
         'image/x-xbitmap'                                                           => 'xbm',
         'image/x-xpixmap'                                                           => 'xpm',
         'image/x-xwindowdump'                                                       => 'xwd',
-        'message/rfc822'                                                            => array('eml', 'mime'),
-        'model/iges'                                                                => array('igs', 'iges'),
-        'model/mesh'                                                                => array('msh', 'mesh', 'silo'),
+        'message/rfc822'                                                            => ['eml', 'mime'],
+        'model/iges'                                                                => ['igs', 'iges'],
+        'model/mesh'                                                                => ['msh', 'mesh', 'silo'],
         'model/vnd.collada+xml'                                                     => 'dae',
         'model/vnd.dwf'                                                             => 'dwf',
         'model/vnd.gdl'                                                             => 'gdl',
         'model/vnd.gtw'                                                             => 'gtw',
         'model/vnd.mts'                                                             => 'mts',
         'model/vnd.vtu'                                                             => 'vtu',
-        'model/vrml'                                                                => array('wrl', 'vrml'),
+        'model/vrml'                                                                => ['wrl', 'vrml'],
         'model/x3d+binary'                                                          => 'x3db',
         'model/x3d+vrml'                                                            => 'x3dv',
         'model/x3d+xml'                                                             => 'x3d',
         'text/cache-manifest'                                                       => 'appcache',
-        'text/calendar'                                                             => array('ics', 'ifb'),
+        'text/calendar'                                                             => ['ics', 'ifb'],
         'text/css'                                                                  => 'css',
         'text/csv'                                                                  => 'csv',
-        'text/html'                                                                 => array('html', 'htm'),
+        'text/html'                                                                 => ['html', 'htm'],
         'text/n3'                                                                   => 'n3',
-        'text/plain'                                                                => array(
+        'text/plain'                                                                => [
             'txt',
             'text',
             'conf',
@@ -457,21 +478,21 @@ class File extends Base
             'list',
             'log',
             'in'
-        ),
+        ],
         'text/prs.lines.tag'                                                        => 'dsc',
         'text/richtext'                                                             => 'rtx',
-        'text/sgml'                                                                 => array('sgml', 'sgm'),
+        'text/sgml'                                                                 => ['sgml', 'sgm'],
         'text/tab-separated-values'                                                 => 'tsv',
-        'text/troff'                                                                => array(
+        'text/troff'                                                                => [
             't',
             'tr',
             'roff',
             'man',
             'me',
             'ms'
-        ),
+        ],
         'text/turtle'                                                               => 'ttl',
-        'text/uri-list'                                                             => array('uri', 'uris', 'urls'),
+        'text/uri-list'                                                             => ['uri', 'uris', 'urls'],
         'text/vcard'                                                                => 'vcard',
         'text/vnd.curl'                                                             => 'curl',
         'text/vnd.curl.dcurl'                                                       => 'dcurl',
@@ -486,11 +507,11 @@ class File extends Base
         'text/vnd.sun.j2me.app-descriptor'                                          => 'jad',
         'text/vnd.wap.wml'                                                          => 'wml',
         'text/vnd.wap.wmlscript'                                                    => 'wmls',
-        'text/x-asm'                                                                => array('s', 'asm'),
-        'text/x-fortran'                                                            => array('f', 'for', 'f77', 'f90'),
+        'text/x-asm'                                                                => ['s', 'asm'],
+        'text/x-fortran'                                                            => ['f', 'for', 'f77', 'f90'],
         'text/x-java-source'                                                        => 'java',
         'text/x-opml'                                                               => 'opml',
-        'text/x-pascal'                                                             => array('p', 'pas'),
+        'text/x-pascal'                                                             => ['p', 'pas'],
         'text/x-nfo'                                                                => 'nfo',
         'text/x-setext'                                                             => 'etx',
         'text/x-sfv'                                                                => 'sfv',
@@ -503,31 +524,31 @@ class File extends Base
         'video/h263'                                                                => 'h263',
         'video/h264'                                                                => 'h264',
         'video/jpeg'                                                                => 'jpgv',
-        'video/jpm'                                                                 => array('jpm', 'jpgm'),
+        'video/jpm'                                                                 => ['jpm', 'jpgm'],
         'video/mj2'                                                                 => 'mj2',
         'video/mp4'                                                                 => 'mp4',
-        'video/mpeg'                                                                => array('mpeg', 'mpg', 'mpe', 'm1v', 'm2v'),
+        'video/mpeg'                                                                => ['mpeg', 'mpg', 'mpe', 'm1v', 'm2v'],
         'video/ogg'                                                                 => 'ogv',
-        'video/quicktime'                                                           => array('qt', 'mov'),
-        'video/vnd.dece.hd'                                                         => array('uvh', 'uvvh'),
-        'video/vnd.dece.mobile'                                                     => array('uvm', 'uvvm'),
-        'video/vnd.dece.pd'                                                         => array('uvp', 'uvvp'),
-        'video/vnd.dece.sd'                                                         => array('uvs', 'uvvs'),
-        'video/vnd.dece.video'                                                      => array('uvv', 'uvvv'),
+        'video/quicktime'                                                           => ['qt', 'mov'],
+        'video/vnd.dece.hd'                                                         => ['uvh', 'uvvh'],
+        'video/vnd.dece.mobile'                                                     => ['uvm', 'uvvm'],
+        'video/vnd.dece.pd'                                                         => ['uvp', 'uvvp'],
+        'video/vnd.dece.sd'                                                         => ['uvs', 'uvvs'],
+        'video/vnd.dece.video'                                                      => ['uvv', 'uvvv'],
         'video/vnd.dvb.file'                                                        => 'dvb',
         'video/vnd.fvt'                                                             => 'fvt',
-        'video/vnd.mpegurl'                                                         => array('mxu', 'm4u'),
+        'video/vnd.mpegurl'                                                         => ['mxu', 'm4u'],
         'video/vnd.ms-playready.media.pyv'                                          => 'pyv',
-        'video/vnd.uvvu.mp4'                                                        => array('uvu', 'uvvu'),
+        'video/vnd.uvvu.mp4'                                                        => ['uvu', 'uvvu'],
         'video/vnd.vivo'                                                            => 'viv',
         'video/webm'                                                                => 'webm',
         'video/x-f4v'                                                               => 'f4v',
         'video/x-fli'                                                               => 'fli',
         'video/x-flv'                                                               => 'flv',
         'video/x-m4v'                                                               => 'm4v',
-        'video/x-matroska'                                                          => array('mkv', 'mk3d', 'mks'),
+        'video/x-matroska'                                                          => ['mkv', 'mk3d', 'mks'],
         'video/x-mng'                                                               => 'mng',
-        'video/x-ms-asf'                                                            => array('asf', 'asx'),
+        'video/x-ms-asf'                                                            => ['asf', 'asx'],
         'video/x-ms-vob'                                                            => 'vob',
         'video/x-ms-wm'                                                             => 'wm',
         'video/x-ms-wmv'                                                            => 'wmv',
@@ -535,7 +556,7 @@ class File extends Base
         'video/x-ms-wvx'                                                            => 'wvx',
         'video/x-msvideo'                                                           => 'avi',
         'video/x-sgi-movie'                                                         => 'movie',
-    );
+    ];
 
     /**
      * Get a random MIME type
@@ -543,7 +564,7 @@ class File extends Base
      * @return string
      * @example 'video/avi'
      */
-    public static function mimeType()
+    public static function mimeType(): string
     {
         return static::randomElement(array_keys(static::$mimeTypes));
     }
@@ -554,7 +575,7 @@ class File extends Base
      * @example avi
      * @return string
      */
-    public static function fileExtension()
+    public static function fileExtension(): string
     {
         $random_extension = static::randomElement(array_values(static::$mimeTypes));
 
@@ -566,30 +587,33 @@ class File extends Base
      *
      * @param  string  $sourceDirectory The directory to look for random file taking
      * @param  string  $targetDirectory
-     * @param  boolean $fullPath        Whether to have the full path or just the filename
+     * @param  bool    $fullPath        Whether to have the full path or just the filename
      * @return string
      */
-    public static function file($sourceDirectory = '/tmp', $targetDirectory = '/tmp', $fullPath = true)
-    {
+    public static function file(
+        string $sourceDirectory = self::TMP_DIR,
+        string $targetDirectory = self::TMP_DIR,
+        bool $fullPath = true
+    ): string {
         if (!is_dir($sourceDirectory)) {
-            throw new \InvalidArgumentException(sprintf('Source directory %s does not exist or is not a directory.', $sourceDirectory));
+            throw new InvalidArgumentException(sprintf('Source directory %s does not exist or is not a directory.', $sourceDirectory));
         }
 
         if (!is_dir($targetDirectory)) {
-            throw new \InvalidArgumentException(sprintf('Target directory %s does not exist or is not a directory.', $targetDirectory));
+            throw new InvalidArgumentException(sprintf('Target directory %s does not exist or is not a directory.', $targetDirectory));
         }
 
         if ($sourceDirectory == $targetDirectory) {
-            throw new \InvalidArgumentException('Source and target directories must differ.');
+            throw new InvalidArgumentException('Source and target directories must differ.');
         }
 
         // Drop . and .. and reset array keys
-        $files = array_filter(array_values(array_diff(scandir($sourceDirectory), array('.', '..'))), function ($file) use ($sourceDirectory) {
+        $files = array_filter(array_values(array_diff(scandir($sourceDirectory), ['.', '..'])), function ($file) use ($sourceDirectory) {
             return is_file($sourceDirectory . DIRECTORY_SEPARATOR . $file) && is_readable($sourceDirectory . DIRECTORY_SEPARATOR . $file);
         });
 
         if (empty($files)) {
-            throw new \InvalidArgumentException(sprintf('Source directory %s is empty.', $sourceDirectory));
+            throw new InvalidArgumentException(sprintf('Source directory %s is empty.', $sourceDirectory));
         }
 
         $sourceFullPath = $sourceDirectory . DIRECTORY_SEPARATOR . static::randomElement($files);
@@ -598,7 +622,7 @@ class File extends Base
         $destinationFullPath = $targetDirectory . DIRECTORY_SEPARATOR . $destinationFile;
 
         if (false === copy($sourceFullPath, $destinationFullPath)) {
-            return false;
+            throw new InvalidArgumentException(sprintf('Copy from source: %s to destination: %s failed', $sourceDirectory, $targetDirectory));
         }
 
         return $fullPath ? $destinationFullPath : $destinationFile;
