@@ -1,16 +1,44 @@
 <?php
+declare(strict_types=1);
 
 namespace Faker\Provider;
+
+use InvalidArgumentException;
+use RuntimeException;
+
+use function is_null;
+use function is_dir;
+use function is_writable;
+use function in_array;
+use function function_exists;
+use function sprintf;
+use function md5;
+use function uniqid;
+use function sys_get_temp_dir;
+use function fopen;
+use function fclose;
+use function unlink;
+use function copy;
+use function ini_get;
+use function curl_init;
+use function curl_setopt;
+use function curl_exec;
+use function curl_getinfo;
+use function curl_close;
 
 /**
  * Depends on image generation from http://lorempixel.com/
  */
 class Image extends Base
 {
-    protected static $categories = array(
+
+    /**
+     * @var string[]
+     */
+    protected static array $categories = [
         'abstract', 'animals', 'business', 'cats', 'city', 'food', 'nightlife',
         'fashion', 'people', 'nature', 'sports', 'technics', 'transport'
-    );
+    ];
 
     /**
      * Generate the URL that will return a random image
@@ -28,8 +56,14 @@ class Image extends Base
      *
      * @return string
      */
-    public static function imageUrl($width = 640, $height = 480, $category = null, $randomize = true, $word = null, $gray = false)
-    {
+    public static function imageUrl(
+        int $width = 640,
+        int $height = 480,
+        ?string $category = null,
+        bool $randomize = true,
+        ?string $word = null,
+        bool $gray = false
+    ): string {
         $baseUrl = "https://lorempixel.com/";
         $url = "{$width}/{$height}/";
 
@@ -39,7 +73,7 @@ class Image extends Base
 
         if ($category) {
             if (!in_array($category, static::$categories)) {
-                throw new \InvalidArgumentException(sprintf('Unknown image category "%s"', $category));
+                throw new InvalidArgumentException(sprintf('Unknown image category "%s"', $category));
             }
             $url .= "{$category}/";
             if ($word) {
@@ -59,14 +93,31 @@ class Image extends Base
      *
      * Requires curl, or allow_url_fopen to be on in php.ini.
      *
+     * @param string|null $dir
+     * @param int $width
+     * @param int $height
+     * @param string|null $category
+     * @param bool $fullPath
+     * @param bool $randomize
+     * @param bool|null $word
+     * @param bool $gray
+     * @return string
      * @example '/path/to/dir/13b73edae8443990be1aa8f1a483bc27.jpg'
      */
-    public static function image($dir = null, $width = 640, $height = 480, $category = null, $fullPath = true, $randomize = true, $word = null, $gray = false)
-    {
+    public static function image(
+        ?string $dir = null,
+        int $width = 640,
+        int $height = 480,
+        ?string $category = null,
+        bool $fullPath = true,
+        bool $randomize = true,
+        bool $word = null,
+        bool $gray = false
+    ): string {
         $dir = is_null($dir) ? sys_get_temp_dir() : $dir; // GNU/Linux / OS X / Windows compatible
         // Validate directory path
         if (!is_dir($dir) || !is_writable($dir)) {
-            throw new \InvalidArgumentException(sprintf('Cannot write to directory "%s"', $dir));
+            throw new InvalidArgumentException(sprintf('Cannot write to directory "%s"', $dir));
         }
 
         // Generate a random filename. Use the server address so that a file
@@ -91,13 +142,12 @@ class Image extends Base
                 unlink($filepath);
 
                 // could not contact the distant URL or HTTP error - fail silently.
-                return false;
+                return '';
             }
         } elseif (ini_get('allow_url_fopen')) {
-            // use remote fopen() via copy()
-            $success = copy($url, $filepath);
+            copy($url, $filepath);
         } else {
-            return new \RuntimeException('The image formatter downloads an image from a remote HTTP server. Therefore, it requires that PHP can request remote hosts, either via cURL or fopen()');
+            throw new RuntimeException('The image formatter downloads an image from a remote HTTP server. Therefore, it requires that PHP can request remote hosts, either via cURL or fopen()');
         }
 
         return $fullPath ? $filepath : $filename;
