@@ -1,27 +1,36 @@
 <?php
+declare(strict_types=1);
 
 namespace Faker\Provider;
 
+use DateTime as DateTime;
 use Faker\Calculator\Iban;
 use Faker\Calculator\Luhn;
 
 class Payment extends Base
 {
-    public static $expirationDateFormat = "m/y";
 
-    protected static $cardVendors = array(
+    /**
+     * @var string
+     */
+    public static string $expirationDateFormat = "m/y";
+
+    /**
+     * @var string[]
+     */
+    protected static array $cardVendors = [
         'Visa', 'Visa', 'Visa', 'Visa', 'Visa',
         'MasterCard', 'MasterCard', 'MasterCard', 'MasterCard', 'MasterCard',
         'American Express', 'Discover Card', 'Visa Retired'
-    );
+    ];
 
     /**
-     * @var array List of card brand masks for generating valid credit card numbers
+     * @var string[][] List of card brand masks for generating valid credit card numbers
      * @see https://en.wikipedia.org/wiki/Payment_card_number Reference for existing prefixes
      * @see https://www.mastercard.us/en-us/issuers/get-support/2-series-bin-expansion.html MasterCard 2017 2-Series BIN Expansion
      */
-    protected static $cardParams = array(
-        'Visa' => array(
+    protected static array $cardParams = [
+        'Visa' => [
             "4539###########",
             "4556###########",
             "4916###########",
@@ -31,8 +40,8 @@ class Payment extends Base
             "4485###########",
             "4716###########",
             "4##############"
-        ),
-        'Visa Retired' => array(
+        ],
+        'Visa Retired' => [
             "4539########",
             "4556########",
             "4916########",
@@ -42,8 +51,8 @@ class Payment extends Base
             "4485########",
             "4716########",
             "4###########",
-        ),
-        'MasterCard' => array(
+        ],
+        'MasterCard' => [
             "2221###########",
             "23#############",
             "24#############",
@@ -55,90 +64,90 @@ class Payment extends Base
             "53#############",
             "54#############",
             "55#############"
-        ),
-        'American Express' => array(
+        ],
+        'American Express' => [
             "34############",
             "37############"
-        ),
-        'Discover Card' => array(
+        ],
+        'Discover Card' => [
             "6011###########"
-        ),
-    );
+        ],
+    ];
 
     /**
      * @var array list of IBAN formats, source: @link https://www.swift.com/standards/data-standards/iban
      */
-    protected static $ibanFormats = array(
-        'AD' => array(array('n', 4),    array('n', 4),  array('c', 12)),
-        'AE' => array(array('n', 3),    array('n', 16)),
-        'AL' => array(array('n', 8),    array('c', 16)),
-        'AT' => array(array('n', 5),    array('n', 11)),
-        'AZ' => array(array('a', 4),    array('c', 20)),
-        'BA' => array(array('n', 3),    array('n', 3),  array('n', 8),  array('n', 2)),
-        'BE' => array(array('n', 3),    array('n', 7),  array('n', 2)),
-        'BG' => array(array('a', 4),    array('n', 4),  array('n', 2),  array('c', 8)),
-        'BH' => array(array('a', 4),    array('c', 14)),
-        'BR' => array(array('n', 8),    array('n', 5),  array('n', 10), array('a', 1),  array('c', 1)),
-        'CH' => array(array('n', 5),    array('c', 12)),
-        'CR' => array(array('n', 4),    array('n', 14)),
-        'CY' => array(array('n', 3),    array('n', 5),  array('c', 16)),
-        'CZ' => array(array('n', 4),    array('n', 6),  array('n', 10)),
-        'DE' => array(array('n', 8),    array('n', 10)),
-        'DK' => array(array('n', 4),    array('n', 9),  array('n', 1)),
-        'DO' => array(array('c', 4),    array('n', 20)),
-        'EE' => array(array('n', 2),    array('n', 2),  array('n', 11), array('n', 1)),
-        'ES' => array(array('n', 4),    array('n', 4),  array('n', 1),  array('n', 1),  array('n', 10)),
-        'FI' => array(array('n', 6),    array('n', 7),  array('n', 1)),
-        'FR' => array(array('n', 5),    array('n', 5),  array('c', 11), array('n', 2)),
-        'GB' => array(array('a', 4),    array('n', 6),  array('n', 8)),
-        'GE' => array(array('a', 2),    array('n', 16)),
-        'GI' => array(array('a', 4),    array('c', 15)),
-        'GR' => array(array('n', 3),    array('n', 4),  array('c', 16)),
-        'GT' => array(array('c', 4),    array('c', 20)),
-        'HR' => array(array('n', 7),    array('n', 10)),
-        'HU' => array(array('n', 3),    array('n', 4),  array('n', 1),  array('n', 15), array('n', 1)),
-        'IE' => array(array('a', 4),    array('n', 6),  array('n', 8)),
-        'IL' => array(array('n', 3),    array('n', 3),  array('n', 13)),
-        'IS' => array(array('n', 4),    array('n', 2),  array('n', 6),  array('n', 10)),
-        'IT' => array(array('a', 1),    array('n', 5),  array('n', 5),  array('c', 12)),
-        'KW' => array(array('a', 4),    array('n', 22)),
-        'KZ' => array(array('n', 3),    array('c', 13)),
-        'LB' => array(array('n', 4),    array('c', 20)),
-        'LI' => array(array('n', 5),    array('c', 12)),
-        'LT' => array(array('n', 5),    array('n', 11)),
-        'LU' => array(array('n', 3),    array('c', 13)),
-        'LV' => array(array('a', 4),    array('c', 13)),
-        'MC' => array(array('n', 5),    array('n', 5),  array('c', 11), array('n', 2)),
-        'MD' => array(array('c', 2),    array('c', 18)),
-        'ME' => array(array('n', 3),    array('n', 13), array('n', 2)),
-        'MK' => array(array('n', 3),    array('c', 10), array('n', 2)),
-        'MR' => array(array('n', 5),    array('n', 5),  array('n', 11), array('n', 2)),
-        'MT' => array(array('a', 4),    array('n', 5),  array('c', 18)),
-        'MU' => array(array('a', 4),    array('n', 2),  array('n', 2),  array('n', 12), array('n', 3),  array('a', 3)),
-        'NL' => array(array('a', 4),    array('n', 10)),
-        'NO' => array(array('n', 4),    array('n', 6),  array('n', 1)),
-        'PK' => array(array('a', 4),    array('c', 16)),
-        'PL' => array(array('n', 8),    array('n', 16)),
-        'PS' => array(array('a', 4),    array('c', 21)),
-        'PT' => array(array('n', 4),    array('n', 4),  array('n', 11), array('n', 2)),
-        'RO' => array(array('a', 4),    array('c', 16)),
-        'RS' => array(array('n', 3),    array('n', 13), array('n', 2)),
-        'SA' => array(array('n', 2),    array('c', 18)),
-        'SE' => array(array('n', 3),    array('n', 16), array('n', 1)),
-        'SI' => array(array('n', 5),    array('n', 8),  array('n', 2)),
-        'SK' => array(array('n', 4),    array('n', 6),  array('n', 10)),
-        'SM' => array(array('a', 1),    array('n', 5),  array('n', 5),  array('c', 12)),
-        'TN' => array(array('n', 2),    array('n', 3),  array('n', 13), array('n', 2)),
-        'TR' => array(array('n', 5),    array('n', 1),  array('c', 16)),
-        'VG' => array(array('a', 4),    array('n', 16)),
-    );
+    protected static array $ibanFormats = [
+        'AD' => [['n', 4],    ['n', 4],  ['c', 12]],
+        'AE' => [['n', 3],    ['n', 16]],
+        'AL' => [['n', 8],    ['c', 16]],
+        'AT' => [['n', 5],    ['n', 11]],
+        'AZ' => [['a', 4],    ['c', 20]],
+        'BA' => [['n', 3],    ['n', 3],  ['n', 8],  ['n', 2]],
+        'BE' => [['n', 3],    ['n', 7],  ['n', 2]],
+        'BG' => [['a', 4],    ['n', 4],  ['n', 2],  ['c', 8]],
+        'BH' => [['a', 4],    ['c', 14]],
+        'BR' => [['n', 8],    ['n', 5],  ['n', 10], ['a', 1],  ['c', 1]],
+        'CH' => [['n', 5],    ['c', 12]],
+        'CR' => [['n', 4],    ['n', 14]],
+        'CY' => [['n', 3],    ['n', 5],  ['c', 16]],
+        'CZ' => [['n', 4],    ['n', 6],  ['n', 10]],
+        'DE' => [['n', 8],    ['n', 10]],
+        'DK' => [['n', 4],    ['n', 9],  ['n', 1]],
+        'DO' => [['c', 4],    ['n', 20]],
+        'EE' => [['n', 2],    ['n', 2],  ['n', 11], ['n', 1]],
+        'ES' => [['n', 4],    ['n', 4],  ['n', 1],  ['n', 1],  ['n', 10]],
+        'FI' => [['n', 6],    ['n', 7],  ['n', 1]],
+        'FR' => [['n', 5],    ['n', 5],  ['c', 11], ['n', 2]],
+        'GB' => [['a', 4],    ['n', 6],  ['n', 8]],
+        'GE' => [['a', 2],    ['n', 16]],
+        'GI' => [['a', 4],    ['c', 15]],
+        'GR' => [['n', 3],    ['n', 4],  ['c', 16]],
+        'GT' => [['c', 4],    ['c', 20]],
+        'HR' => [['n', 7],    ['n', 10]],
+        'HU' => [['n', 3],    ['n', 4],  ['n', 1],  ['n', 15], ['n', 1]],
+        'IE' => [['a', 4],    ['n', 6],  ['n', 8]],
+        'IL' => [['n', 3],    ['n', 3],  ['n', 13]],
+        'IS' => [['n', 4],    ['n', 2],  ['n', 6],  ['n', 10]],
+        'IT' => [['a', 1],    ['n', 5],  ['n', 5],  ['c', 12]],
+        'KW' => [['a', 4],    ['n', 22]],
+        'KZ' => [['n', 3],    ['c', 13]],
+        'LB' => [['n', 4],    ['c', 20]],
+        'LI' => [['n', 5],    ['c', 12]],
+        'LT' => [['n', 5],    ['n', 11]],
+        'LU' => [['n', 3],    ['c', 13]],
+        'LV' => [['a', 4],    ['c', 13]],
+        'MC' => [['n', 5],    ['n', 5],  ['c', 11], ['n', 2]],
+        'MD' => [['c', 2],    ['c', 18]],
+        'ME' => [['n', 3],    ['n', 13], ['n', 2]],
+        'MK' => [['n', 3],    ['c', 10], ['n', 2]],
+        'MR' => [['n', 5],    ['n', 5],  ['n', 11], ['n', 2]],
+        'MT' => [['a', 4],    ['n', 5],  ['c', 18]],
+        'MU' => [['a', 4],    ['n', 2],  ['n', 2],  ['n', 12], ['n', 3],  ['a', 3]],
+        'NL' => [['a', 4],    ['n', 10]],
+        'NO' => [['n', 4],    ['n', 6],  ['n', 1]],
+        'PK' => [['a', 4],    ['c', 16]],
+        'PL' => [['n', 8],    ['n', 16]],
+        'PS' => [['a', 4],    ['c', 21]],
+        'PT' => [['n', 4],    ['n', 4],  ['n', 11], ['n', 2]],
+        'RO' => [['a', 4],    ['c', 16]],
+        'RS' => [['n', 3],    ['n', 13], ['n', 2]],
+        'SA' => [['n', 2],    ['c', 18]],
+        'SE' => [['n', 3],    ['n', 16], ['n', 1]],
+        'SI' => [['n', 5],    ['n', 8],  ['n', 2]],
+        'SK' => [['n', 4],    ['n', 6],  ['n', 10]],
+        'SM' => [['a', 1],    ['n', 5],  ['n', 5],  ['c', 12]],
+        'TN' => [['n', 2],    ['n', 3],  ['n', 13], ['n', 2]],
+        'TR' => [['n', 5],    ['n', 1],  ['c', 16]],
+        'VG' => [['a', 4],    ['n', 16]],
+    ];
 
     /**
      * @return string Returns a credit card vendor name
      *
      * @example 'MasterCard'
      */
-    public static function creditCardType()
+    public static function creditCardType(): string
     {
         return static::randomElement(static::$cardVendors);
     }
@@ -146,14 +155,14 @@ class Payment extends Base
     /**
      * Returns the String of a credit card number.
      *
-     * @param string  $type      Supporting any of 'Visa', 'MasterCard', 'American Express', and 'Discover'
-     * @param boolean $formatted Set to true if the output string should contain one separator every 4 digits
+     * @param string|null  $type     Supporting any of 'Visa', 'MasterCard', 'American Express', and 'Discover'
+     * @param bool         $formatted Set to true if the output string should contain one separator every 4 digits
      * @param string  $separator Separator string for formatting card number. Defaults to dash (-).
      * @return string
      *
      * @example '4485480221084675'
      */
-    public static function creditCardNumber($type = null, $formatted = false, $separator = '-')
+    public static function creditCardNumber(?string $type = null, bool $formatted = false, string $separator = '-'): string
     {
         if (is_null($type)) {
             $type = static::creditCardType();
@@ -175,11 +184,11 @@ class Payment extends Base
     }
 
     /**
-     * @param boolean $valid True (by default) to get a valid expiration date, false to get a maybe valid date
-     * @return \DateTime
+     * @param bool $valid True (by default) to get a valid expiration date, false to get a maybe valid date
+     * @return DateTime
      * @example 04/13
      */
-    public function creditCardExpirationDate($valid = true)
+    public function creditCardExpirationDate(bool $valid = true): DateTime
     {
         if ($valid) {
             return $this->generator->dateTimeBetween('now', '36 months');
@@ -189,42 +198,42 @@ class Payment extends Base
     }
 
     /**
-     * @param boolean $valid                True (by default) to get a valid expiration date, false to get a maybe valid date
-     * @param string  $expirationDateFormat
+     * @param bool $valid                True (by default) to get a valid expiration date, false to get a maybe valid date
+     * @param string|null  $expirationDateFormat
      * @return string
      * @example '04/13'
      */
-    public function creditCardExpirationDateString($valid = true, $expirationDateFormat = null)
+    public function creditCardExpirationDateString(bool $valid = true, ?string $expirationDateFormat = null): string
     {
         return $this->creditCardExpirationDate($valid)->format(is_null($expirationDateFormat) ? static::$expirationDateFormat : $expirationDateFormat);
     }
 
     /**
-     * @param  boolean $valid True (by default) to get a valid expiration date, false to get a maybe valid date
+     * @param  bool $valid True (by default) to get a valid expiration date, false to get a maybe valid date
      * @return array
      */
-    public function creditCardDetails($valid = true)
+    public function creditCardDetails(bool $valid = true): array
     {
         $type = static::creditCardType();
 
-        return array(
+        return [
             'type'   => $type,
             'number' => static::creditCardNumber($type),
             'name'   => $this->generator->name(),
             'expirationDate' => $this->creditCardExpirationDateString($valid)
-        );
+        ];
     }
 
     /**
      * International Bank Account Number (IBAN)
      *
      * @link http://en.wikipedia.org/wiki/International_Bank_Account_Number
-     * @param  string  $countryCode ISO 3166-1 alpha-2 country code
-     * @param  string  $prefix      for generating bank account number of a specific bank
-     * @param  integer $length      total length without country code and 2 check digits
+     * @param  string|null  $countryCode ISO 3166-1 alpha-2 country code
+     * @param  string       $prefix      for generating bank account number of a specific bank
+     * @param  int|null     $length      total length without country code and 2 check digits
      * @return string
      */
-    public static function iban($countryCode = null, $prefix = '', $length = null)
+    public static function iban(?string $countryCode = null, string $prefix = '', ?int $length = null): string
     {
         $countryCode = is_null($countryCode) ? self::randomKey(self::$ibanFormats) : strtoupper($countryCode);
 
@@ -279,7 +288,7 @@ class Payment extends Base
      * @link    http://en.wikipedia.org/wiki/ISO_9362
      * @return  string Swift/Bic number
      */
-    public static function swiftBicNumber()
+    public static function swiftBicNumber(): string
     {
         return self::regexify("^([A-Z]){4}([A-Z]){2}([0-9A-Z]){2}([0-9A-Z]{3})?$");
     }
