@@ -1,13 +1,17 @@
 <?php
+declare(strict_types=1);
 
 namespace Faker\Provider\en_US;
+
+use function join;
 
 class PhoneNumber extends \Faker\Provider\PhoneNumber
 {
     /**
      * @see https://en.wikipedia.org/wiki/National_conventions_for_writing_telephone_numbers#United_States.2C_Canada.2C_and_other_NANP_countries
+     * @var string[]
      */
-    protected static $formats = array(
+    protected static array $formats = [
         // International format
         '+1-{{areaCode}}-{{exchangeCode}}-####',
         '+1 ({{areaCode}}) {{exchangeCode}}-####',
@@ -41,31 +45,41 @@ class PhoneNumber extends \Faker\Provider\PhoneNumber
         '({{areaCode}}) {{exchangeCode}}-#### x#####',
         '1-{{areaCode}}-{{exchangeCode}}-#### x#####',
         '{{areaCode}}.{{exchangeCode}}.#### x#####'
-    );
+    ];
 
     /**
      * @see https://en.wikipedia.org/wiki/Toll-free_telephone_number#United_States
+     * @var int[]
      */
-    protected static $tollFreeAreaCodes = array(
+    protected static array $tollFreeAreaCodes = [
         800, 844, 855, 866, 877, 888
-    );
-    protected static $tollFreeFormats = array(
+    ];
+
+    /**
+     * @var string[]
+     */
+    protected static array $tollFreeFormats = [
         // Standard formats
         '{{tollFreeAreaCode}}-{{exchangeCode}}-####',
         '({{tollFreeAreaCode}}) {{exchangeCode}}-####',
         '1-{{tollFreeAreaCode}}-{{exchangeCode}}-####',
         '{{tollFreeAreaCode}}.{{exchangeCode}}.####',
-    );
+    ];
 
-    public function tollFreeAreaCode()
+    /**
+     * @return int
+     */
+    public function tollFreeAreaCode(): int
     {
         return self::randomElement(static::$tollFreeAreaCodes);
     }
 
-    public function tollFreePhoneNumber()
+    /**
+     * @return string
+     */
+    public function tollFreePhoneNumber(): string
     {
         $format = self::randomElement(static::$tollFreeFormats);
-
         return self::numerify($this->generator->parse($format));
     }
 
@@ -76,7 +90,7 @@ class PhoneNumber extends \Faker\Provider\PhoneNumber
      *
      * @return string
      */
-    public static function areaCode()
+    public static function areaCode(): string
     {
         $digits[] = self::numberBetween(2, 9);
         $digits[] = self::randomDigit();
@@ -92,7 +106,7 @@ class PhoneNumber extends \Faker\Provider\PhoneNumber
      *
      * @return string
      */
-    public static function exchangeCode()
+    public static function exchangeCode(): string
     {
         $digits[] = self::numberBetween(2, 9);
         $digits[] = self::randomDigit();
