@@ -1,25 +1,39 @@
 <?php
+declare(strict_types=1);
 
 namespace Faker;
+
+use OverflowException;
+
+use Faker\Api\FakerGeneratorInterface;
+use Faker\Api\FakerProviderInterface;
+
+use function array_key_exists;
+use function call_user_func_array;
+use function serialize;
+use function sprintf;
 
 /**
  * Proxy for other generators, to return only unique values. Works with
  * Faker\Generator\Base->unique()
  */
-class UniqueGenerator
+class UniqueGenerator implements FakerGeneratorInterface
 {
-    protected $generator;
-    protected $maxRetries;
-    protected $uniques = array();
+
+    /**
+     * @var array
+     */
+    protected array $uniques;
 
     /**
      * @param Generator $generator
      * @param integer $maxRetries
      */
-    public function __construct(Generator $generator, $maxRetries = 10000)
-    {
-        $this->generator = $generator;
-        $this->maxRetries = $maxRetries;
+    public function __construct(
+        protected FakerGeneratorInterface $generator,
+        protected int $maxRetries = 10000
+    ) {
+        $this->uniques = [];
     }
 
     /**
@@ -27,9 +41,9 @@ class UniqueGenerator
      * @param string $attribute
      * @return mixed
      */
-    public function __get($attribute)
+    public function __get(string $attribute): mixed
     {
-        return $this->__call($attribute, array());
+        return $this->__call($attribute, []);
     }
 
     /**
@@ -38,21 +52,35 @@ class UniqueGenerator
      * @param array $arguments
      * @return mixed
      */
-    public function __call($name, $arguments)
+    public function __call(string $name, array $arguments): mixed
     {
         if (!isset($this->uniques[$name])) {
-            $this->uniques[$name] = array();
+            $this->uniques[$name] = [];
         }
         $i = 0;
         do {
-            $res = call_user_func_array(array($this->generator, $name), $arguments);
+            $res = call_user_func_array([$this->generator, $name], $arguments);
             $i++;
             if ($i > $this->maxRetries) {
-                throw new \OverflowException(sprintf('Maximum retries of %d reached without finding a unique value', $this->maxRetries));
+                throw new OverflowException(sprintf('Maximum retries of %d reached without finding a unique value', $this->maxRetries));
             }
         } while (array_key_exists(serialize($res), $this->uniques[$name]));
         $this->uniques[$name][serialize($res)]= null;
 
         return $res;
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function addProvider(FakerProviderInterface $provider): void
+    {}
+
+    /**
+     * @inheritDoc
+     */
+    public function getProviders(): array
+    {
+        return [];
     }
 }
