@@ -1,37 +1,45 @@
 <?php
+declare(strict_types=1);
 
 namespace Faker\Calculator;
+
+use Faker\Api\FakerCalculatorInterface;
+
+use function strlen;
+use function preg_match;
+use function intval;
+use function substr;
 
 /**
  * Utility class for validating EAN-8 and EAN-13 numbers
  *
  * @package Faker\Calculator
  */
-class Ean
+class Ean implements FakerCalculatorInterface
 {
     /** @var string EAN validation pattern */
-    const PATTERN = '/^(?:\d{8}|\d{13})$/';
+    public const PATTERN = '/^(?:\d{8}|\d{13})$/';
 
     /**
      * Computes the checksum of an EAN number.
      *
      * @see https://en.wikipedia.org/wiki/International_Article_Number
      *
-     * @param string $digits
+     * @param string $value
      * @return int
      */
-    public static function checksum($digits)
+    public static function checksum(string $value): int
     {
-        $length = strlen($digits);
+        $length = strlen($value);
 
         $even = 0;
         for ($i = $length - 1; $i >= 0; $i -= 2) {
-            $even += $digits[$i];
+            $even += (int) $value[$i];
         }
 
         $odd = 0;
         for ($i = $length - 2; $i >= 0; $i -= 2) {
-            $odd += $digits[$i];
+            $odd += (int) $value[$i];
         }
 
         return (10 - ((3 * $even + $odd) % 10)) % 10;
@@ -41,15 +49,15 @@ class Ean
      * Checks whether the provided number is an EAN compliant number and that
      * the checksum is correct.
      *
-     * @param string $ean An EAN number
-     * @return boolean
+     * @param string $value An EAN number
+     * @return bool
      */
-    public static function isValid($ean)
+    public static function isValid(string $value): bool
     {
-        if (!preg_match(self::PATTERN, $ean)) {
+        if (!preg_match(self::PATTERN, $value)) {
             return false;
         }
 
-        return self::checksum(substr($ean, 0, -1)) === intval(substr($ean, -1));
+        return self::checksum(substr($value, 0, -1)) === intval(substr($value, -1));
     }
 }
