@@ -1,6 +1,12 @@
 <?php
+declare(strict_types=1);
 
 namespace Faker\Provider;
+
+use function mt_rand;
+use function mt_getrandmax;
+use function call_user_func;
+use function floor;
 
 class Biased extends Base
 {
@@ -11,15 +17,15 @@ class Biased extends Base
      * The algorithm creates two doubles, x ∈ [0, 1], y ∈ [0, 1) and checks whether the
      * return value of $function for x is greater than or equal to y. If this is
      * the case the number is accepted and x is mapped to the appropriate integer
-     * between $min and $max. Otherwise two new doubles are created until the pair
+     * between $min and $max. Otherwise, two new doubles are created until the pair
      * is accepted.
      *
-     * @param integer $min Minimum value of the generated integers.
-     * @param integer $max Maximum value of the generated integers.
-     * @param callable $function A function mapping x ∈ [0, 1] onto a double ∈ [0, 1]
-     * @return integer An integer between $min and $max.
+     * @param int $min Minimum value of the generated integers.
+     * @param int $max Maximum value of the generated integers.
+     * @param callable|string $function A function mapping x ∈ [0, 1] onto a double ∈ [0, 1]
+     * @return int An integer between $min and $max.
      */
-    public function biasedNumberBetween($min = 0, $max = 100, $function = 'sqrt')
+    public function biasedNumberBetween(int $min = 0, int $max = 100, callable|string $function = 'sqrt'): int
     {
         do {
             $x = mt_rand() / mt_getrandmax();
@@ -33,9 +39,9 @@ class Biased extends Base
      * 'unbiased' creates an unbiased distribution by giving
      * each value the same value of one.
      *
-     * @return integer
+     * @return int
      */
-    protected static function unbiased()
+    protected static function unbiased(): int
     {
         return 1;
     }
@@ -44,9 +50,10 @@ class Biased extends Base
      * 'linearLow' favors lower numbers. The probability decreases
      * in a linear fashion.
      *
-     * @return integer
+     * @param int|float $x
+     * @return int|float
      */
-    protected static function linearLow($x)
+    protected static function linearLow(int|float $x): int|float
     {
         return 1 - $x;
     }
@@ -54,11 +61,11 @@ class Biased extends Base
     /**
      * 'linearHigh' favors higher numbers. The probability increases
      * in a linear fashion.
-     *
-     * @return integer
+     * @param int|float $x
+     * @return int|float
      */
-    protected static function linearHigh($x)
+    protected static function linearHigh(int|float $x): int|float
     {
-        return $x;
+        return $x + 1;
     }
 }
