@@ -10,7 +10,7 @@ final class AddressTest extends TestCase
 {
     private $faker;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $faker = new Generator();
         $faker->addProvider(new Address($faker));
@@ -19,16 +19,16 @@ final class AddressTest extends TestCase
 
     public function testLatitude()
     {
-        $latitude = $this->faker->latitude();
-        $this->assertInternalType('float', $latitude);
+        $latitude = $this->faker->latitude;
+        $this->assertIsFloat($latitude);
         $this->assertGreaterThanOrEqual(-90, $latitude);
         $this->assertLessThanOrEqual(90, $latitude);
     }
 
     public function testLongitude()
     {
-        $longitude = $this->faker->longitude();
-        $this->assertInternalType('float', $longitude);
+        $longitude = $this->faker->longitude;
+        $this->assertIsFloat($longitude);
         $this->assertGreaterThanOrEqual(-180, $longitude);
         $this->assertLessThanOrEqual(180, $longitude);
     }
@@ -36,11 +36,11 @@ final class AddressTest extends TestCase
     public function testCoordinate()
     {
         $coordinate = $this->faker->localCoordinates();
-        $this->assertInternalType('array', $coordinate);
-        $this->assertInternalType('float', $coordinate['latitude']);
+        $this->assertIsArray($coordinate);
+        $this->assertIsFloat($coordinate['latitude']);
         $this->assertGreaterThanOrEqual(-90, $coordinate['latitude']);
         $this->assertLessThanOrEqual(90, $coordinate['latitude']);
-        $this->assertInternalType('float', $coordinate['longitude']);
+        $this->assertIsFloat($coordinate['longitude']);
         $this->assertGreaterThanOrEqual(-180, $coordinate['longitude']);
         $this->assertLessThanOrEqual(180, $coordinate['longitude']);
     }
