@@ -10,7 +10,7 @@ final class BarcodeTest extends TestCase
 {
     private $faker;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $faker = new Generator();
         $faker->addProvider(new Barcode($faker));
@@ -21,7 +21,7 @@ final class BarcodeTest extends TestCase
     public function testEan8()
     {
         $code = $this->faker->ean8();
-        $this->assertRegExp('/^\d{8}$/i', $code);
+        $this->assertMatchesRegularExpression('/^\d{8}$/i', $code);
         $codeWithoutChecksum = substr($code, 0, -1);
         $checksum = substr($code, -1);
         $this->assertEquals(TestableBarcode::eanChecksum($codeWithoutChecksum), $checksum);
@@ -30,7 +30,7 @@ final class BarcodeTest extends TestCase
     public function testEan13()
     {
         $code = $this->faker->ean13();
-        $this->assertRegExp('/^\d{13}$/i', $code);
+        $this->assertMatchesRegularExpression('/^\d{13}$/i', $code);
         $codeWithoutChecksum = substr($code, 0, -1);
         $checksum = substr($code, -1);
         $this->assertEquals(TestableBarcode::eanChecksum($codeWithoutChecksum), $checksum);
@@ -39,7 +39,7 @@ final class BarcodeTest extends TestCase
 
 final class TestableBarcode extends Barcode
 {
-    public static function eanChecksum($input)
+    public static function eanChecksum(string $input): int
     {
         return parent::eanChecksum($input);
     }
