@@ -12,7 +12,7 @@ final class BiasedTest extends TestCase
     protected $generator;
     protected $results = array();
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->generator = new Generator();
         $this->generator->addProvider(new Biased($this->generator));
@@ -39,36 +39,6 @@ final class BiasedTest extends TestCase
             $assumed /= 1;
             $this->assertGreaterThan(self::NUMBERS * $assumed * .95, $amount, "Value was more than 5 percent under the expected value");
             $this->assertLessThan(self::NUMBERS * $assumed * 1.05, $amount, "Value was more than 5 percent over the expected value");
-        }
-    }
-
-    public function testLinearHigh()
-    {
-        $this->performFake(array('\Faker\Provider\Biased', 'linearHigh'));
-
-        foreach ($this->results as $number => $amount) {
-            // integral
-            $assumed = 0.5 * pow(1 / self::MAX * $number, 2) - 0.5 * pow(1 / self::MAX * ($number - 1), 2);
-            // calculate the fraction of the whole area
-            $assumed /= pow(1, 2) * .5;
-            $this->assertGreaterThan(self::NUMBERS * $assumed * .9, $amount, "Value was more than 10 percent under the expected value");
-            $this->assertLessThan(self::NUMBERS * $assumed * 1.1, $amount, "Value was more than 10 percent over the expected value");
-        }
-    }
-
-    public function testLinearLow()
-    {
-        $this->performFake(array('\Faker\Provider\Biased', 'linearLow'));
-
-        foreach ($this->results as $number => $amount) {
-            // integral
-            $assumed = -0.5 * pow(1 / self::MAX * $number, 2) - -0.5 * pow(1 / self::MAX * ($number - 1), 2);
-            // shift the graph up
-            $assumed += 1 / self::MAX;
-            // calculate the fraction of the whole area
-            $assumed /= pow(1, 2) * .5;
-            $this->assertGreaterThan(self::NUMBERS * $assumed * .9, $amount, "Value was more than 10 percent under the expected value");
-            $this->assertLessThan(self::NUMBERS * $assumed * 1.1, $amount, "Value was more than 10 percent over the expected value");
         }
     }
 }
