@@ -14,7 +14,7 @@ final class CompanyTest extends TestCase
      */
     private $faker;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $faker = new Generator();
         $faker->addProvider(new Company($faker));
@@ -26,6 +26,6 @@ final class CompanyTest extends TestCase
     {
         $jobTitle = $this->faker->jobTitle();
         $pattern = '/^[A-Za-z]+$/';
-        $this->assertRegExp($pattern, $jobTitle);
+        $this->assertMatchesRegularExpression($pattern, $jobTitle);
     }
 }
