@@ -7,13 +7,13 @@ use PHPUnit\Framework\TestCase;
 
 final class DateTimeTest extends TestCase
 {
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->defaultTz = 'UTC';
         DateTimeProvider::setDefaultTimezone($this->defaultTz);
     }
 
-    protected function tearDown()
+    protected function tearDown(): void
     {
         DateTimeProvider::setDefaultTimezone();
     }
@@ -71,7 +71,7 @@ final class DateTimeTest extends TestCase
     public function testUnixTime()
     {
         $timestamp = DateTimeProvider::unixTime();
-        $this->assertInternalType('int', $timestamp);
+        $this->assertIsInt($timestamp);
         $this->assertGreaterThanOrEqual(0, $timestamp);
         $this->assertLessThanOrEqual(time(), $timestamp);
     }
@@ -163,7 +163,7 @@ final class DateTimeTest extends TestCase
     public function testIso8601()
     {
         $date = DateTimeProvider::iso8601();
-        $this->assertRegExp('/^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}[+-Z](\d{4})?$/', $date);
+        $this->assertMatchesRegularExpression('/^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}[+-Z](\d{4})?$/', $date);
         $this->assertGreaterThanOrEqual(new \DateTime('@0'), new \DateTime($date));
         $this->assertLessThanOrEqual(new \DateTime(), new \DateTime($date));
     }
@@ -171,7 +171,7 @@ final class DateTimeTest extends TestCase
     public function testDate()
     {
         $date = DateTimeProvider::date();
-        $this->assertRegExp('/^\d{4}-\d{2}-\d{2}$/', $date);
+        $this->assertMatchesRegularExpression('/^\d{4}-\d{2}-\d{2}$/', $date);
         $this->assertGreaterThanOrEqual(new \DateTime('@0'), new \DateTime($date));
         $this->assertLessThanOrEqual(new \DateTime(), new \DateTime($date));
     }
@@ -179,7 +179,7 @@ final class DateTimeTest extends TestCase
     public function testTime()
     {
         $date = DateTimeProvider::time();
-        $this->assertRegExp('/^\d{2}:\d{2}:\d{2}$/', $date);
+        $this->assertMatchesRegularExpression('/^\d{2}:\d{2}:\d{2}$/', $date);
     }
 
     /**
@@ -198,8 +198,8 @@ final class DateTimeTest extends TestCase
     public function providerDateTimeBetween()
     {
         return array(
-            array('-1 year', false),
-            array('-1 year', null),
+            array('-1 year', '+2 hours'),
+            array('-1 year', '+1 day'),
             array('-1 day', '-1 hour'),
             array('-1 day', 'now'),
         );
@@ -208,8 +208,9 @@ final class DateTimeTest extends TestCase
     /**
      *
      * @dataProvider providerDateTimeInInterval
+     * @return void
      */
-    public function testDateTimeInInterval($start, $interval = "+5 days", $isInFuture)
+    public function testDateTimeInInterval(string $start, string  $interval , bool $isInFuture): void
     {
         $date = DateTimeProvider::dateTimeInInterval($start, $interval);
         $this->assertInstanceOf('\DateTime', $date);
