@@ -3,28 +3,29 @@
 namespace Faker\Test\Provider;
 
 use Faker\Provider\Image;
+use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
 
 final class ImageTest extends TestCase
 {
     public function testImageUrlUses640x680AsTheDefaultSize()
     {
-        $this->assertRegExp('#^https://lorempixel.com/640/480/#', Image::imageUrl());
+        $this->assertMatchesRegularExpression('#^https://lorempixel.com/640/480/#', Image::imageUrl());
     }
 
     public function testImageUrlAcceptsCustomWidthAndHeight()
     {
-        $this->assertRegExp('#^https://lorempixel.com/800/400/#', Image::imageUrl(800, 400));
+        $this->assertMatchesRegularExpression('#^https://lorempixel.com/800/400/#', Image::imageUrl(800, 400));
     }
 
     public function testImageUrlAcceptsCustomCategory()
     {
-        $this->assertRegExp('#^https://lorempixel.com/800/400/nature/#', Image::imageUrl(800, 400, 'nature'));
+        $this->assertMatchesRegularExpression('#^https://lorempixel.com/800/400/nature/#', Image::imageUrl(800, 400, 'nature'));
     }
 
     public function testImageUrlAcceptsCustomText()
     {
-        $this->assertRegExp('#^https://lorempixel.com/800/400/nature/Faker#', Image::imageUrl(800, 400, 'nature', false, 'Faker'));
+        $this->assertMatchesRegularExpression('#^https://lorempixel.com/800/400/nature/Faker#', Image::imageUrl(800, 400, 'nature', false, 'Faker'));
     }
 
     public function testImageUrlReturnsLinkToRegularImageWhenGrayIsFalse()
@@ -60,15 +61,17 @@ final class ImageTest extends TestCase
         $url = Image::imageUrl(800, 400);
         $splitUrl = preg_split('/\?/', $url);
 
-        $this->assertEquals(count($splitUrl), 2);
-        $this->assertRegexp('#\d{5}#', $splitUrl[1]);
+        $this->assertCount(2, $splitUrl);
+        $this->assertMatchesRegularExpression('#\d{5}#', $splitUrl[1]);
     }
 
+
     /**
-     * @expectedException \InvalidArgumentException
+     * @return void
      */
-    public function testUrlWithDimensionsAndBadCategory()
+    public function testUrlWithDimensionsAndBadCategory(): void
     {
+        $this->expectException(InvalidArgumentException::class);
         Image::imageUrl(800, 400, 'bullhonky');
     }
 
