@@ -16,7 +16,7 @@ final class InternetTest extends TestCase
      */
     private $faker;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $faker = new Generator();
         $faker->addProvider(new Lorem($faker));
@@ -52,7 +52,7 @@ final class InternetTest extends TestCase
         $this->loadLocalProviders($locale);
         $pattern = '/^(?!(?:(?:\\x22?\\x5C[\\x00-\\x7E]\\x22?)|(?:\\x22?[^\\x5C\\x22]\\x22?)){255,})(?!(?:(?:\\x22?\\x5C[\\x00-\\x7E]\\x22?)|(?:\\x22?[^\\x5C\\x22]\\x22?)){65,}@)(?:(?:[\\x21\\x23-\\x27\\x2A\\x2B\\x2D\\x2F-\\x39\\x3D\\x3F\\x5E-\\x7E]+)|(?:\\x22(?:[\\x01-\\x08\\x0B\\x0C\\x0E-\\x1F\\x21\\x23-\\x5B\\x5D-\\x7F]|(?:\\x5C[\\x00-\\x7F]))*\\x22))(?:\\.(?:(?:[\\x21\\x23-\\x27\\x2A\\x2B\\x2D\\x2F-\\x39\\x3D\\x3F\\x5E-\\x7E]+)|(?:\\x22(?:[\\x01-\\x08\\x0B\\x0C\\x0E-\\x1F\\x21\\x23-\\x5B\\x5D-\\x7F]|(?:\\x5C[\\x00-\\x7F]))*\\x22)))*@(?:(?:(?!.*[^.]{64,})(?:(?:(?:xn--)?[a-z0-9]+(?:-+[a-z0-9]+)*\\.){1,126}){1,}(?:(?:[a-z][a-z0-9]*)|(?:(?:xn--)[a-z0-9]+))(?:-+[a-z0-9]+)*)|(?:\\[(?:(?:IPv6:(?:(?:[a-f0-9]{1,4}(?::[a-f0-9]{1,4}){7})|(?:(?!(?:.*[a-f0-9][:\\]]){7,})(?:[a-f0-9]{1,4}(?::[a-f0-9]{1,4}){0,5})?::(?:[a-f0-9]{1,4}(?::[a-f0-9]{1,4}){0,5})?)))|(?:(?:IPv6:(?:(?:[a-f0-9]{1,4}(?::[a-f0-9]{1,4}){5}:)|(?:(?!(?:.*[a-f0-9]:){5,})(?:[a-f0-9]{1,4}(?::[a-f0-9]{1,4}){0,3})?::(?:[a-f0-9]{1,4}(?::[a-f0-9]{1,4}){0,3}:)?)))?(?:(?:25[0-5])|(?:2[0-4][0-9])|(?:1[0-9]{2})|(?:[1-9]?[0-9]))(?:\\.(?:(?:25[0-5])|(?:2[0-4][0-9])|(?:1[0-9]{2})|(?:[1-9]?[0-9]))){3}))\\]))$/iD';
         $emailAddress = $this->faker->email();
-        $this->assertRegExp($pattern, $emailAddress);
+        $this->assertMatchesRegularExpression($pattern, $emailAddress);
     }
 
     /**
@@ -67,7 +67,7 @@ final class InternetTest extends TestCase
         $this->loadLocalProviders($locale);
         $pattern = '/^[A-Za-z0-9]+([._][A-Za-z0-9]+)*$/';
         $username = $this->faker->username();
-        $this->assertRegExp($pattern, $username);
+        $this->assertMatchesRegularExpression($pattern, $username);
     }
 
     /**
@@ -82,7 +82,7 @@ final class InternetTest extends TestCase
         $this->loadLocalProviders($locale);
         $pattern = '/^[a-z]+(\.[a-z]+)+$/';
         $domainName = $this->faker->domainName();
-        $this->assertRegExp($pattern, $domainName);
+        $this->assertMatchesRegularExpression($pattern, $domainName);
     }
 
     /**
@@ -97,7 +97,7 @@ final class InternetTest extends TestCase
         $this->loadLocalProviders($locale);
         $pattern = '/^[a-z]+$/';
         $domainWord = $this->faker->domainWord();
-        $this->assertRegExp($pattern, $domainWord);
+        $this->assertMatchesRegularExpression($pattern, $domainWord);
     }
 
     public function loadLocalProviders($locale)
@@ -119,7 +119,7 @@ final class InternetTest extends TestCase
 
     public function testPasswordIsValid()
     {
-        $this->assertRegexp('/^.{6}$/', $this->faker->password(6, 6));
+        $this->assertMatchesRegularExpression('/^.{6}$/', $this->faker->password(6, 6));
     }
 
     public function testSlugIsValid()
@@ -131,7 +131,7 @@ final class InternetTest extends TestCase
 
     public function testUrlIsValid()
     {
-        $url = $this->faker->url();
+        $url = $this->faker->url;
         $this->assertNotFalse(filter_var($url, FILTER_VALIDATE_URL));
     }
 
@@ -142,26 +142,26 @@ final class InternetTest extends TestCase
 
     public function testIpv4()
     {
-        $this->assertNotFalse(filter_var($this->faker->ipv4(), FILTER_VALIDATE_IP, FILTER_FLAG_IPV4));
+        $this->assertNotFalse(filter_var($this->faker->ipv4, FILTER_VALIDATE_IP, FILTER_FLAG_IPV4));
     }
 
     public function testIpv4NotLocalNetwork()
     {
-        $this->assertNotRegExp('/\A0\./', $this->faker->ipv4());
+        $this->assertDoesNotMatchRegularExpression('/\A0\./', $this->faker->ipv4);
     }
 
     public function testIpv4NotBroadcast()
     {
-        $this->assertNotEquals('255.255.255.255', $this->faker->ipv4());
+        $this->assertNotEquals('255.255.255.255', $this->faker->ipv4);
     }
 
     public function testIpv6()
     {
-        $this->assertNotFalse(filter_var($this->faker->ipv6(), FILTER_VALIDATE_IP, FILTER_FLAG_IPV6));
+        $this->assertNotFalse(filter_var($this->faker->ipv6, FILTER_VALIDATE_IP, FILTER_FLAG_IPV6));
     }
 
     public function testMacAddress()
     {
-        $this->assertRegExp('/^([0-9A-F]{2}[:]){5}([0-9A-F]{2})$/i', Internet::macAddress());
+        $this->assertMatchesRegularExpression('/^([0-9A-F]{2}[:]){5}([0-9A-F]{2})$/i', Internet::macAddress());
     }
 }
