@@ -9,13 +9,14 @@ use Faker\Provider\Base as BaseProvider;
 use Faker\Provider\DateTime as DateTimeProvider;
 use Faker\Provider\Payment as PaymentProvider;
 use Faker\Provider\Person as PersonProvider;
+use http\Exception\InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
 
 final class PaymentTest extends TestCase
 {
     private $faker;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $faker = new Generator();
         $faker->addProvider(new BaseProvider($faker));
@@ -67,13 +68,13 @@ final class PaymentTest extends TestCase
     public function testCreditCardNumberReturnsValidCreditCardNumber($type, $regexp)
     {
         $cardNumber = $this->faker->creditCardNumber($type);
-        $this->assertRegExp($regexp, $cardNumber);
+        $this->assertMatchesRegularExpression($regexp, $cardNumber);
         $this->assertTrue(Luhn::isValid($cardNumber));
     }
 
     public function testCreditCardNumberCanFormatOutput()
     {
-        $this->assertRegExp('/^6011-\d{4}-\d{4}-\d{4}$/', $this->faker->creditCardNumber('Discover Card', true));
+        $this->assertMatchesRegularExpression('/^6011-\d{4}-\d{4}-\d{4}$/', $this->faker->creditCardNumber('Discover Card', true));
     }
 
     public function testCreditCardExpirationDateReturnsValidDateByDefault()
@@ -165,6 +166,7 @@ final class PaymentTest extends TestCase
 
         if (!isset($this->ibanFormats[$countryCode])) {
             // No IBAN format available
+            $this->markTestSkipped("No IBAN format available for $countryCode");
             return;
         }
 
@@ -172,6 +174,9 @@ final class PaymentTest extends TestCase
 
         try {
             $iban = $this->faker->bankAccountNumber;
+            if (!$iban) {
+                throw new InvalidArgumentException('bankAccountNumber not implemented for');
+            }
         } catch (\InvalidArgumentException $e) {
             // Not implemented, nothing to test
             $this->markTestSkipped("bankAccountNumber not implemented for $locale");
@@ -179,7 +184,7 @@ final class PaymentTest extends TestCase
         }
 
         // Test format
-        $this->assertRegExp($this->ibanFormats[$countryCode], $iban);
+        $this->assertMatchesRegularExpression($this->ibanFormats[$countryCode], $iban);
 
         // Test checksum
         $this->assertTrue(Iban::isValid($iban), "Checksum for $iban is invalid");
@@ -201,7 +206,7 @@ final class PaymentTest extends TestCase
         $iban = $this->faker->iban($countryCode);
 
         // Test format
-        $this->assertRegExp($regex, $iban);
+        $this->assertMatchesRegularExpression($regex, $iban);
 
         // Test checksum
         $this->assertTrue(Iban::isValid($iban), "Checksum for $iban is invalid");
