@@ -37,10 +37,10 @@ final class ProviderOverrideTest extends TestCase
     {
         $faker = Faker\Factory::create($locale);
 
-        $this->assertRegExp(static::TEST_STRING_REGEX, $faker->city);
-        $this->assertRegExp(static::TEST_STRING_REGEX, $faker->postcode);
-        $this->assertRegExp(static::TEST_STRING_REGEX, $faker->address);
-        $this->assertRegExp(static::TEST_STRING_REGEX, $faker->country);
+        $this->assertMatchesRegularExpression(static::TEST_STRING_REGEX, $faker->city);
+        $this->assertMatchesRegularExpression(static::TEST_STRING_REGEX, $faker->postcode);
+        $this->assertMatchesRegularExpression(static::TEST_STRING_REGEX, $faker->address);
+        $this->assertMatchesRegularExpression(static::TEST_STRING_REGEX, $faker->country);
     }
 
 
@@ -52,7 +52,7 @@ final class ProviderOverrideTest extends TestCase
     {
         $faker = Faker\Factory::create($locale);
 
-        $this->assertRegExp(static::TEST_STRING_REGEX, $faker->company);
+        $this->assertMatchesRegularExpression(static::TEST_STRING_REGEX, $faker->company);
     }
 
 
@@ -64,8 +64,8 @@ final class ProviderOverrideTest extends TestCase
     {
         $faker = Faker\Factory::create($locale);
 
-        $this->assertRegExp(static::TEST_STRING_REGEX, $faker->century);
-        $this->assertRegExp(static::TEST_STRING_REGEX, $faker->timezone);
+        $this->assertMatchesRegularExpression(static::TEST_STRING_REGEX, $faker->century);
+        $this->assertMatchesRegularExpression(static::TEST_STRING_REGEX, $faker->timezone);
     }
 
 
@@ -81,12 +81,12 @@ final class ProviderOverrideTest extends TestCase
 
         $faker = Faker\Factory::create($locale);
 
-        $this->assertRegExp(static::TEST_STRING_REGEX, $faker->userName);
+        $this->assertMatchesRegularExpression(static::TEST_STRING_REGEX, $faker->userName);
 
-        $this->assertRegExp(static::TEST_EMAIL_REGEX, $faker->email);
-        $this->assertRegExp(static::TEST_EMAIL_REGEX, $faker->safeEmail);
-        $this->assertRegExp(static::TEST_EMAIL_REGEX, $faker->freeEmail);
-        $this->assertRegExp(static::TEST_EMAIL_REGEX, $faker->companyEmail);
+        $this->assertMatchesRegularExpression(static::TEST_EMAIL_REGEX, $faker->email);
+        $this->assertMatchesRegularExpression(static::TEST_EMAIL_REGEX, $faker->safeEmail);
+        $this->assertMatchesRegularExpression(static::TEST_EMAIL_REGEX, $faker->freeEmail);
+        $this->assertMatchesRegularExpression(static::TEST_EMAIL_REGEX, $faker->companyEmail);
     }
 
 
@@ -98,10 +98,10 @@ final class ProviderOverrideTest extends TestCase
     {
         $faker = Faker\Factory::create($locale);
 
-        $this->assertRegExp(static::TEST_STRING_REGEX, $faker->name);
-        $this->assertRegExp(static::TEST_STRING_REGEX, $faker->title);
-        $this->assertRegExp(static::TEST_STRING_REGEX, $faker->firstName);
-        $this->assertRegExp(static::TEST_STRING_REGEX, $faker->lastName);
+        $this->assertMatchesRegularExpression(static::TEST_STRING_REGEX, $faker->name);
+        $this->assertMatchesRegularExpression(static::TEST_STRING_REGEX, $faker->title);
+        $this->assertMatchesRegularExpression(static::TEST_STRING_REGEX, $faker->firstName);
+        $this->assertMatchesRegularExpression(static::TEST_STRING_REGEX, $faker->lastName);
     }
 
 
@@ -113,7 +113,7 @@ final class ProviderOverrideTest extends TestCase
     {
         $faker = Faker\Factory::create($locale);
 
-        $this->assertRegExp(static::TEST_STRING_REGEX, $faker->phoneNumber);
+        $this->assertMatchesRegularExpression(static::TEST_STRING_REGEX, $faker->phoneNumber);
     }
 
 
@@ -125,7 +125,7 @@ final class ProviderOverrideTest extends TestCase
     {
         $faker = Faker\Factory::create($locale);
 
-        $this->assertRegExp(static::TEST_STRING_REGEX, $faker->userAgent);
+        $this->assertMatchesRegularExpression(static::TEST_STRING_REGEX, $faker->userAgent);
     }
 
 
@@ -139,7 +139,7 @@ final class ProviderOverrideTest extends TestCase
     {
         $faker = Faker\Factory::create($locale);
 
-        $this->assertRegExp(static::TEST_STRING_REGEX, $faker->uuid);
+        $this->assertMatchesRegularExpression(static::TEST_STRING_REGEX, $faker->uuid);
     }
 
 
